streamlit-app/requirements.txt
streamlit
requests
pandas
plotly
anthropic
python-dotenv
Pillow